fixed : redundancies in files
