+++ packages/formstr-app/src/App.js
@@ -3,9 +3,7 @@ import "./App.css";
 import Routing from "./components/Routing";
 import { ProfileProvider } from "./provider/ProfileProvider";
 import { ApplicationProvider } from "./provider/ApplicationProvider";
-import { UserProvider } from "./provider/UserProvider";
 import { ListProvider } from "./provider/ListProvider";
-import { AppContextProvider } from "./provider/AppProvider";
 
 function App() {
   return (
@@ -13,13 +11,9 @@ function App() {
       <div className="App">
         <ProfileProvider>
           <ApplicationProvider>
-            <AppContextProvider>
-              <UserProvider>
-                <ListProvider>
-                  <Routing />
-                </ListProvider>
-              </UserProvider>
-            </AppContextProvider>
+            <ListProvider>
+              <Routing />
+            </ListProvider>
           </ApplicationProvider>
         </ProfileProvider>
       </div>
