--- packages/formstr-app/src/hooks/useUserContext/index.ts
@@ -1,15 +1,0 @@
-import { useContext } from "react";
-import { UserContext } from "../../provider/UserProvider";
-
-
-export function useUserContext() {
-    const context = useContext(UserContext);
-  
-    if (!context) {
-      throw new Error("UserContext must be used within a UserContextProvider");
-    }
-  
-    return context;
-}
-
-export default UserContext;
